We now proceed with delete only if entity is found. 	Change on 2013/01/12 by psimakov <psimakov@google.com> ------------- Created by MOE: http://code.google.com/p/moe-java MOE_MIGRATED_REVID=41065422

+++ coursebuilder/modules/admin/config.py
@@ -173,14 +173,16 @@ class ConfigPropertyEditor(object):
         # Delete if exists.
         try:
             entity = config.ConfigPropertyEntity.get_by_key_name(name)
-            old_value = entity.value
-            entity.delete()
+            if entity:
+                old_value = entity.value
+                entity.delete()
+
+                models.EventEntity.record(
+                    'delete-property', users.get_current_user(), json.dumps({
+                        'name': name, 'value': str(old_value)}))
+
         except db.BadKeyError:
             pass
-
-        models.EventEntity.record(
-            'delete-property', users.get_current_user(), json.dumps({
-                'name': name, 'value': str(old_value)}))
 
         self.redirect('/admin?action=settings')
